Add files via upload
This is a revenue model I've constructed that has 2 different revenue sources within a company that feeds into the main revenue model. The 2 revenue sources are vastly different, one being lead generation that are sold to partnered agents. The 2nd is a subscription based service offered through the company to our partnered agents as well.